third_party/blink/web_tests/fullscreen/full-screen-min-max-width-height-video-expected.txt

EVENT(webkitfullscreenchange)
EXPECTED (computedStyle.minWidth == '0px') OK
EXPECTED (computedStyle.maxWidth == 'none') OK
EXPECTED (computedStyle.minHeight == '0px') OK
EXPECTED (computedStyle.maxHeight == 'none') OK
EXPECTED (computedStyle.width == '800px') OK
EXPECTED (computedStyle.height == '600px') OK
END OF TEST
